requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

streamlit==1.26.0
tensorflow-intel==2.13.0
